+++ svd-encoder/src/config.rs
@@ -166,6 +166,26 @@ impl FromStr for FieldBitRangeFormat {
     }
 }
 
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub enum Sorting {
+    Offset,
+    OffsetReversed,
+    Name,
+}
+
+impl FromStr for Sorting {
+    type Err = ();
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "Offset" => Ok(Self::Offset),
+            "OffsetReversed" => Ok(Self::OffsetReversed),
+            "Name" => Ok(Self::OffsetReversed),
+            _ => Err(()),
+        }
+    }
+}
+
 #[derive(Clone, Copy, Debug)]
 #[non_exhaustive]
 /// Advanced encoder options
@@ -183,6 +203,9 @@ pub struct Config {
     ///
     /// format: hex, dec
     pub peripheral_base_address: NumberFormat,
+
+    /// Sort peripherals in specified order
+    pub peripheral_sorting: Option<Sorting>,
 
     /// Format of addressBlock's offset element
     ///
@@ -208,6 +231,9 @@ pub struct Config {
     ///
     /// format: hex, dec
     pub cluster_address_offset: NumberFormat,
+
+    /// Sort registers and clusters in specified order
+    pub register_cluster_sorting: Option<Sorting>,
 
     /// Format of register's name-kind elements
     /// - `derivedFrom`
@@ -245,6 +271,9 @@ pub struct Config {
     /// `None` means keep the original bitRange
     pub field_bit_range: Option<FieldBitRangeFormat>,
 
+    /// Sort fields in specified order
+    pub field_sorting: Option<Sorting>,
+
     /// Format of enumeratedValues's name-kind elements
     /// - `derivedFrom`
     /// - `name`
@@ -278,11 +307,13 @@ impl Default for Config {
         Self {
             peripheral_name: None,
             peripheral_base_address: NumberFormat::UpperHex8,
+            peripheral_sorting: None,
             address_block_offset: NumberFormat::UpperHex,
             address_block_size: NumberFormat::UpperHex,
             interrupt_name: None,
             cluster_name: None,
             cluster_address_offset: NumberFormat::UpperHex,
+            register_cluster_sorting: None,
             register_name: None,
             register_address_offset: NumberFormat::UpperHex,
             register_size: NumberFormat::LowerHex,
@@ -290,6 +321,7 @@ impl Default for Config {
             register_reset_mask: NumberFormat::UpperHex16,
             field_name: None,
             field_bit_range: None,
+            field_sorting: None,
             enumerated_values_name: None,
             enumerated_value_name: None,
             enumerated_value_value: NumberFormat::Dec,
@@ -308,11 +340,15 @@ impl Config {
         match name {
             "peripheral_name" => self.peripheral_name = Some(value.parse().unwrap()),
             "peripheral_base_address" => self.peripheral_base_address = value.parse().unwrap(),
+            "peripheral_sorting" => self.peripheral_sorting = Some(value.parse().unwrap()),
             "address_block_offset" => self.address_block_offset = value.parse().unwrap(),
             "address_block_size" => self.address_block_size = value.parse().unwrap(),
             "interrupt_name" => self.interrupt_name = Some(value.parse().unwrap()),
             "cluster_name" => self.cluster_name = Some(value.parse().unwrap()),
             "cluster_address_offset" => self.cluster_address_offset = value.parse().unwrap(),
+            "register_cluster_sorting" => {
+                self.register_cluster_sorting = Some(value.parse().unwrap())
+            }
             "register_name" => self.register_name = Some(value.parse().unwrap()),
             "register_address_offset" => self.register_address_offset = value.parse().unwrap(),
             "register_size" => self.register_size = value.parse().unwrap(),
@@ -320,6 +356,7 @@ impl Config {
             "register_reset_mask" => self.register_reset_mask = value.parse().unwrap(),
             "field_name" => self.field_name = Some(value.parse().unwrap()),
             "field_bit_range" => self.field_bit_range = Some(value.parse().unwrap()),
+            "field_sorting" => self.field_sorting = Some(value.parse().unwrap()),
             "enumerated_values_name" => self.enumerated_values_name = Some(value.parse().unwrap()),
             "enumerated_value_name" => self.enumerated_value_name = Some(value.parse().unwrap()),
             "enumerated_value_value" => self.enumerated_value_value = value.parse().unwrap(),
@@ -345,6 +382,14 @@ impl Config {
     /// format: hex, dec
     pub fn peripheral_base_address(mut self, val: NumberFormat) -> Self {
         self.peripheral_base_address = val;
+        self
+    }
+
+    /// Sort peripherals in specified order
+    ///
+    /// `None` means keep the original order
+    pub fn peripheral_sorting(mut self, val: Option<Sorting>) -> Self {
+        self.peripheral_sorting = val;
         self
     }
 
@@ -381,6 +426,14 @@ impl Config {
     /// format: hex, dec
     pub fn cluster_address_offset(mut self, val: NumberFormat) -> Self {
         self.cluster_address_offset = val;
+        self
+    }
+
+    /// Sort registers and clusters in specified order
+    ///
+    /// `None` means keep the original order
+    pub fn register_cluster_sorting(mut self, val: Option<Sorting>) -> Self {
+        self.register_cluster_sorting = val;
         self
     }
 
@@ -433,6 +486,14 @@ impl Config {
     /// `None` means keep the original bitRange
     pub fn field_bit_range(mut self, val: Option<FieldBitRangeFormat>) -> Self {
         self.field_bit_range = val;
+        self
+    }
+
+    /// Sort fields in specified order
+    ///
+    /// `None` means keep the original order
+    pub fn field_sorting(mut self, val: Option<Sorting>) -> Self {
+        self.field_sorting = val;
         self
     }