+++ svd-encoder/src/peripheral.rs
@@ -3,7 +3,7 @@ use super::{
 };
 
 use crate::{
-    config::{change_case, format_number},
+    config::{change_case, format_number, Sorting},
     svd::{Peripheral, PeripheralInfo},
 };
 
@@ -102,10 +102,24 @@ impl Encode for PeripheralInfo {
         elem.children.append(&mut interrupts?);
 
         if let Some(v) = &self.registers {
-            let children: Result<Vec<_>, _> = v
-                .iter()
-                .map(|e| e.encode_node_with_config(config))
-                .collect();
+            let children: Result<Vec<_>, _> = if let Some(sorting) = config.register_cluster_sorting
+            {
+                let mut refs = v.iter().collect::<Vec<_>>();
+                match sorting {
+                    Sorting::Offset => refs.sort_by_key(|r| r.address_offset()),
+                    Sorting::OffsetReversed => {
+                        refs.sort_by_key(|r| -(r.address_offset() as i32));
+                    }
+                    Sorting::Name => refs.sort_by_key(|r| r.name()),
+                }
+                refs.into_iter()
+                    .map(|e| e.encode_node_with_config(config))
+                    .collect()
+            } else {
+                v.iter()
+                    .map(|e| e.encode_node_with_config(config))
+                    .collect()
+            };
 
             elem.children.push({
                 let mut e = Element::new("registers");